更新电压估计电源电量 优化电量查找方线性插值法

+++ G030/Core/Src/bsp_tim.c
@@ -8,7 +8,7 @@
 
 /* Private variables ---------------------------------------------------------*/
 static TIM_HandleTypeDef *s_htim = NULL;
-static uint16_t data_time_counter = 0;
+static uint8_t data_time_counter = 0;
 
 /* Exported functions --------------------------------------------------------*/
